refactor: rename configuration module to config
